% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padjust.R
\name{adjust}
\alias{adjust}
\title{Adjusting of combine p-values using null model}
\usage{
adjust(combp, combp_ref)
}
\arguments{
\item{combp}{combined p-value matrix obtaing using gene expression matrix}

\item{combp_ref}{combined p-value matrix obtaing using null model}
}
\value{
A list of three dataframes, One is adjusted p-value datafrmae, second is raw p-value data frame and third is log tranformed p-value data frame
}
\description{
Adjusting of combine p-values using null model
}
\examples{
adjust()
}
