% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvals.R
\name{binorm}
\alias{binorm}
\title{Conversion of non-zero gene FPKM value into p-value using each cells mean and standard deviation}
\usage{
binorm(x)
}
\arguments{
\item{x}{gene expression matrix}
}
\value{
P-value matrix
}
\description{
Conversion of non-zero gene FPKM value into p-value using each cells mean and standard deviation
}
\examples{
binorm()
}
