% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_p_values.R
\name{combine}
\alias{combine}
\title{Combining P values}
\usage{
combine(gene_file, expression_matrix, gnames, Pval1)
}
\arguments{
\item{gene_file}{Pre compiled pathway annotation file}

\item{expression_matrix}{gene expression matrix}

\item{gnames}{rownames of gene expression matrix}

\item{Pval1}{P-value matrix obtained using binorm}
}
\value{
Combined P-value matrix
}
\description{
Combining P values
}
\examples{
combine()
}
