% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayCooccurrence.R
\name{cooccurrence}
\alias{cooccurrence}
\title{Pathway co occurrence}
\usage{
cooccurrence(nulldata, PathwayScores)
}
\arguments{
\item{nulldata}{Pathway score matrix corresponding to null model}

\item{PathwayScores}{Adjusted raw p-values matrix}
}
\value{
A list containing matrices of p-values and correlation for pathway pairs in particular cell type
}
\description{
Pathway co occurrence
}
\examples{
coocurrence()
}
