% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Counttofpkm.R
\name{Counttofpkm}
\alias{Counttofpkm}
\title{Title Conversion of read count data to fpkm}
\usage{
Counttofpkm(countMatrix, length)
}
\arguments{
\item{countMatrix}{Raw read count matrix}

\item{length}{length of genes. Order of length of genes and row names of count data should be same}
}
\value{
fpkm converted matrix
}
\description{
Title Conversion of read count data to fpkm
}
\examples{
Counttofpkm()
}
