% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difcoccur.R
\name{difcoccur}
\alias{difcoccur}
\title{Differential coenriched pathways}
\usage{
difcoccur(data, group)
}
\arguments{
\item{data}{Raw p-value pathway score matrix}

\item{group}{group 1 for cells of interest and rest in group 2}
}
\value{
A list containing matrix of p-values and difference among the groups
}
\description{
Differential coenriched pathways
}
\examples{
difcoccur()
}
