% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distclust.R
\name{dist_clust}
\alias{dist_clust}
\title{Hierarchal clustering of pathway score matrix}
\usage{
dist_clust(pathwayscores, n)
}
\arguments{
\item{pathwayscores}{Raw adjusted p-value matrix}

\item{n}{Number of clusters required for pseudo temporal ordering}
}
\value{
Distance matrix and Clusters
}
\description{
Hierarchal clustering of pathway score matrix
}
\examples{
dist_clust()
}
