% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkdist.R
\name{distance}
\alias{distance}
\title{Two level shrinkage of distance matrix based on nearest neighbour indices and belongingness of cells to same class}
\usage{
distance(dist, class, clusters)
}
\arguments{
\item{dist}{Distance matrix used for hierarchal clustering}

\item{class}{Matrix with number of times cells in same class have top k neighbors in other classes}

\item{clusters}{clusters obtained from hierarchal clustering}
}
\value{
shrinked distance matrix
}
\description{
Two level shrinkage of distance matrix based on nearest neighbour indices and belongingness of cells to same class
}
\examples{
distance()
}
