% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drimpute.R
\name{drimpute}
\alias{drimpute}
\title{Imputation of scATAC-seq profiles}
\usage{
drimpute(countFile)
}
\arguments{
\item{countFile}{scATAC-seq count matrix}
}
\value{
imputed count matrix
}
\description{
Imputation of scATAC-seq profiles
}
\examples{
drimpute()
}
