% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_accessibility_scores.R
\name{global_access}
\alias{global_access}
\title{Global accessibility score calculation}
\usage{
global_access(testfile, referencefile, globalaccess_scores)
}
\arguments{
\item{testfile}{3 column peak file of genomic coordinates for test data}

\item{referencefile}{3 column peak file of genomic coordinates for reference data}

\item{globalaccess_scores}{pre-calculated global accessibility scores}
}
\value{
global accessibility scores for test data
}
\description{
Global accessibility score calculation
}
\examples{
global_access()
}
