% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Showing gradient or continuum of pathways on minimum spanning tree}
\usage{
gradient(pathwayfile, term)
}
\arguments{
\item{pathwayfile}{adjusted p-value matrix}

\item{term}{Pathway term for which gradient needs to be plotted}
}
\value{
gradient of colors for specific pathway term
}
\description{
Showing gradient or continuum of pathways on minimum spanning tree
}
\examples{
gradient()
}
