% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Finding K nearest neighbour}
\usage{
index(pathwayscores, k = 5)
}
\arguments{
\item{pathwayscores}{Raw adjusted p-value matrix}

\item{k}{Number of k nearest neighbour for each sample or cell}
}
\value{
matrix of indices of nearest neighbour
}
\description{
Finding K nearest neighbour
}
\examples{

index()
}
