% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{KNN}
\alias{KNN}
\title{Getting cluster numbers for each of the nearest neighbor of a cell}
\usage{
KNN(pathwayscores, index, clusters)
}
\arguments{
\item{pathwayscores}{Raw adjusted p-value matrix}

\item{index}{indices of nearest neighbours obtained from index function}

\item{clusters}{Clusters obtained from hierarchal clustering}
}
\value{
Matrix having cluster or class number for each of the top nearest neighbor of individual cell
}
\description{
Getting cluster numbers for each of the nearest neighbor of a cell
}
\examples{
KNN()
}
