% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makecolor.R
\name{makecolor}
\alias{makecolor}
\title{Creating gradient of colors for showing continuum of co-occurrence of pathways on minimum spanning tree}
\usage{
makecolor(score1, score2)
}
\arguments{
\item{score1}{scores of first pathway term}

\item{score2}{scores of second pathway term}
}
\value{
matrix of gradient of colors for two pathway terms to be visualized on minimum spanning tree
}
\description{
Creating gradient of colors for showing continuum of co-occurrence of pathways on minimum spanning tree
}
\examples{
makecolor()
}
