% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum_spanning_tree.R
\name{minimum_spanning_tree}
\alias{minimum_spanning_tree}
\title{Finding minimum spanning tree using shrinked distance matrix}
\usage{
minimum_spanning_tree(distance)
}
\arguments{
\item{distance}{shrinked distance matrix}
}
\value{
Minimum spanning tree
}
\description{
Finding minimum spanning tree using shrinked distance matrix
}
\examples{
minimum_spanning_tree()
}
