% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst_plot.R
\name{mst.plot.mod}
\alias{mst.plot.mod}
\title{Plotting minimum spanning tree}
\usage{
mst.plot.mod(x, layout.function = NULL, colors = NULL,
  mst.edge.col = "white", vertex.color = "skyblue", tkplot = FALSE,
  expression = NULL, v.size = FALSE, e.size = FALSE,
  mst.e.size = 1, edge.col.wt = NULL, v.lab = FALSE, e.lab = NULL,
  bg = "black", v.lab.cex = 0.5, e.lab.cex = 0.5,
  v.lab.col = "blue", lab.dist = 0, e.lab.col = "blue", v.sf = c(3,
  12), sf = 0, e.arrow = 0.2, layout.overall = NULL)
}
\arguments{
\item{layout.overall}{}
}
\value{
mst plot
}
\description{
Plotting minimum spanning tree
}
