% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_gene.R
\name{nearest_gene}
\alias{nearest_gene}
\title{Generation of foreground file}
\usage{
nearest_gene(arg1, arg2, arg3, arg4)
}
\arguments{
\item{arg1}{genomic coordinate file}

\item{arg2}{human reference genome file}

\item{arg3}{output file}
}
\value{
foreground file
}
\description{
Generation of foreground file
}
\examples{
nearest_gene()
}
