% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGO.R
\name{runGO}
\alias{runGO}
\title{Calculating pathway enrichment scores for scATAC-seq profiles}
\usage{
runGO(gmtFile, BGfile, countFile, method, globalaccessibility_scores,
  FGfile, promoters = FALSE)
}
\arguments{
\item{gmtFile}{gene-set file}

\item{BGfile}{background file}

\item{countFile}{scATAC-seq count matrix}

\item{method}{If method is chosen as 1, data normalization is performed using global accessibility scores. If selected method is 2 then local accessibility score-based normalization is performed}

\item{globalaccessibility_scores}{global accessibility scores for input count data matrix}

\item{FGfile}{foreground file}

\item{promoters}{whether promoters to be used or not for conversion of scATAC-seq profiles to pathway scores. Default is false}
}
\value{
A list containing two dataframes, One dataframe contains p-values based on hypergeometric test and other one is p-values based on binomial test
}
\description{
Calculating pathway enrichment scores for scATAC-seq profiles
}
\examples{
runGO()
}
