% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teomporal.R
\name{temporaldif}
\alias{temporaldif}
\title{Differential Pathways}
\usage{
temporaldif(data, group)
}
\arguments{
\item{data}{adjusted raw p-value matrix}

\item{group}{group of cell types among which differential pathway analysis needs to be performed.}
}
\value{
A list containing p-value based on wilcoxon rank sum test, fold change based on mean and median
}
\description{
Differential Pathways
}
\examples{
temporaldif()
}
