% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padjust.R
\name{adjust}
\alias{adjust}
\title{Adjusting of combine p-values using null model}
\usage{
adjust(combp, combp_ref)
}
\arguments{
\item{combp}{combined p-value matrix obtained using gene expression matrix}

\item{combp_ref}{combined p-value matrix obtained using null model}
}
\value{
A list of 3 matrices, One is absolute p-value matrix, second is raw adjusted p-value matrix and third is log transformed adjusted p-value matrix
}
\description{
Adjusting of combine p-values using null model
}
\examples{
adjust()
}
