% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{class1}
\alias{class1}
\title{Finding how many times nearest neighbor of cells in each class are belonging to different cluster or class.}
\usage{
class1(clusters, KNN, symmtery = FALSE)
}
\arguments{
\item{KNN}{Matrix obtained from KNN function with cluster number for top nearest neighbors for each of the cell}

\item{Clusters}{clusters obtained from hierarchical clustering}
}
\value{
Matrix with number of times cells in same class have top k neighbors in other classes
}
\description{
Finding how many times nearest neighbor of cells in each class are belonging to different cluster or class.
}
\examples{
class1()
}
