% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_p_values.R
\name{combine}
\alias{combine}
\title{Combining P values}
\usage{
combine(gene_file, expression_matrix, gnames, Pval1, thr = 2)
}
\arguments{
\item{gene_file}{pathway annotation file/gene-set file}

\item{expression_matrix}{gene expression matrix}

\item{gnames}{gene names of expression matrix}

\item{Pval1}{P-value matrix obtained using binorm}

\item{thr}{Based on threshold provided, those gene sets having number of genes greater than the threshold value will to be considered for covariance matrix calculation and combining of p values}
}
\value{
Combined P-value matrix
}
\description{
Combining P values
}
\examples{
combine()
}
