% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difcoccur.R
\name{difcoccur}
\alias{difcoccur}
\title{Differential cooccurrence pathway analysis}
\usage{
difcoccur(data, group)
}
\arguments{
\item{data}{Adjusted raw p-values matrix}

\item{group}{group 1 for cells of interest and rest in group 2}
}
\value{
A list containing matrices of p-values and difference in pathway pairs among the groups
}
\description{
Differential cooccurrence pathway analysis
}
\examples{
difcoccur()
}
