% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Creating gradient of colors for showing continuum of single pathway on minimum spanning tree}
\usage{
gradient(pathwayfile, term)
}
\arguments{
\item{pathwayfile}{Adjusted p-values matrix}

\item{term}{Pathway term for which gradient needs to be plotted}
}
\value{
gradient of colors for specific pathway term
}
\description{
Creating gradient of colors for showing continuum of single pathway on minimum spanning tree
}
\examples{
gradient()
}
